♻️ refactor: redisson관련 repository의 interface를 조금 더 높은 수준으로 추상화
