+++ src/main/java/com/project/cheerha/common/redis/redisson/RedissonLockRepository.java
@@ -1,7 +1,7 @@
 package com.project.cheerha.common.redis.redisson;
 
+import com.project.cheerha.common.scheduler.repository.LockRepository;
 import lombok.RequiredArgsConstructor;
-import org.redisson.api.RBucket;
 import org.redisson.api.RLock;
 import org.redisson.api.RedissonClient;
 import org.springframework.stereotype.Repository;
@@ -10,21 +10,9 @@ import java.util.concurrent.TimeUnit;
 
 @Repository
 @RequiredArgsConstructor
-public class RedissonRepositoryImpl implements RedissonRepository {
+public class RedissonLockRepository implements LockRepository {
 
     private final RedissonClient redissonClient;
-
-    @Override
-    public String getValue(String key) {
-        RBucket<String> bucket = redissonClient.getBucket(key);
-        return bucket.get();
-    }
-
-    @Override
-    public void setValue(String key, String value) {
-        RBucket<String> bucket = redissonClient.getBucket(key);
-        bucket.set(value);
-    }
 
     @Override
     public boolean tryLock(String lockKey, long waitTime, long leaseTime, TimeUnit unit) throws InterruptedException {
