+++ src/main/java/com/project/cheerha/common/scheduler/core/InstanceManager.java
@@ -1,7 +1,7 @@
 package com.project.cheerha.common.scheduler.core;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.project.cheerha.common.redis.redisson.RedissonRepository;
+import com.project.cheerha.common.scheduler.repository.KeyValueRepository;
 import com.project.cheerha.common.util.HealthCheckUtil;
 import com.project.cheerha.common.util.InstanceUtil;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +15,7 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class InstanceManager {
 
-    private final RedissonRepository redissonRepository;
+    private final KeyValueRepository keyValueRepository;
     private final ObjectMapper objectMapper;
     private static final String LATEST_INSTANCE_KEY = "scheduler:latest-instance";
 
@@ -28,7 +28,7 @@ public class InstanceManager {
             return false;
         }
         try {
-            String latestInstanceJson = redissonRepository.getValue(LATEST_INSTANCE_KEY);
+            String latestInstanceJson = keyValueRepository.getValue(LATEST_INSTANCE_KEY);
             if (latestInstanceJson == null) {
                 return true;
             }
@@ -51,7 +51,7 @@ public class InstanceManager {
     public void updateLatestInstance() {
         try {
             //Redis 에서 최신 인스턴스 정보 가져오기
-            String latestInstanceJson = redissonRepository.getValue(LATEST_INSTANCE_KEY);
+            String latestInstanceJson = keyValueRepository.getValue(LATEST_INSTANCE_KEY);
             long currentStartTime = InstanceUtil.getInstanceStartTime().toEpochMilli();
             String currentInstanceId = InstanceUtil.getInstanceId();
 
@@ -71,7 +71,7 @@ public class InstanceManager {
                     "startTime", String.valueOf(currentStartTime)
             );
 
-            redissonRepository.setValue(LATEST_INSTANCE_KEY, objectMapper.writeValueAsString(newLatestInstance));
+            keyValueRepository.setValue(LATEST_INSTANCE_KEY, objectMapper.writeValueAsString(newLatestInstance));
             log.info("최신 인스턴스로 등록됨: {}", newLatestInstance);
         } catch (Exception e) {
             log.error("최신 인스턴스 등록 중 오류 발생", e);
