+++ src/main/java/com/project/cheerha/common/scheduler/producer/TaskRegister.java
@@ -1,7 +1,7 @@
 package com.project.cheerha.common.scheduler.producer;
 
+import com.project.cheerha.common.scheduler.repository.LockRepository;
 import com.project.cheerha.common.scheduler.core.TaskHandler;
-import com.project.cheerha.common.redis.redisson.RedissonRepository;
 import com.project.cheerha.common.scheduler.repository.TaskRepository;
 import com.project.cheerha.common.scheduler.strategy.FixedIntervalStrategy;
 import com.project.cheerha.common.scheduler.strategy.ScheduleStrategy;
@@ -16,7 +16,7 @@ import java.util.concurrent.TimeUnit;
 @RequiredArgsConstructor
 public class TaskRegister {
 
-    private final RedissonRepository redissonRepository;
+    private final LockRepository lockRepository;
     private final TaskRepository taskRepository;
     private final TaskProducer taskProducer;
 
@@ -32,7 +32,7 @@ public class TaskRegister {
         String lastTimeKey = "scheduler:lastScheduledTime:" + taskType;
 
         try {
-            if (redissonRepository.tryLock(lockKey, Math.min(2000, scheduleIntervalMillis / 2), scheduleIntervalMillis / 2, TimeUnit.MILLISECONDS)) {
+            if (lockRepository.tryLock(lockKey, Math.min(2000, scheduleIntervalMillis / 2), scheduleIntervalMillis / 2, TimeUnit.MILLISECONDS)) {
                 try {
                     ScheduleStrategy strategy = handler.getScheduleStrategy();
                     Instant now = Instant.now();
@@ -54,7 +54,7 @@ public class TaskRegister {
                         }
                     }
                 } finally {
-                    redissonRepository.unlock(lockKey);
+                    lockRepository.unlock(lockKey);
                 }
             }
         } catch (InterruptedException e) {
