+++ src/main/java/com/project/cheerha/common/scheduler/repository/KeyValueRepository.java
@@ -1,0 +1,8 @@
+package com.project.cheerha.common.scheduler.repository;
+
+public interface KeyValueRepository {
+
+    //Key-Value
+    String getValue(String key);
+    void setValue(String key, String value);
+}
