+++ src/main/java/com/project/cheerha/common/scheduler/repository/LockRepository.java
@@ -1,12 +1,8 @@
-package com.project.cheerha.common.redis.redisson;
+package com.project.cheerha.common.scheduler.repository;
 
 import java.util.concurrent.TimeUnit;
 
-public interface RedissonRepository {
-
-    //Key-Value
-    String getValue(String key);
-    void setValue(String key, String value);
+public interface LockRepository {
 
     //Lock
     boolean tryLock(String lockKey, long waitTime, long leaseTime, TimeUnit unit) throws InterruptedException;
